move agenda to subdir
