+++ lib/configuratie.include.php
@@ -122,7 +122,7 @@ require_once 'model/framework/DynamicEntityModel.class.php';
 require_once 'model/framework/CachedPersistenceModel.abstract.php';
 require_once 'model/DebugLogModel.class.php';
 require_once 'model/TimerModel.class.php';
-require_once 'model/entity/Agendeerbaar.interface.php';
+require_once 'model/entity/agenda/Agendeerbaar.interface.php';
 require_once 'model/security/AccessModel.class.php';
 require_once 'model/LidInstellingenModel.class.php';
 require_once 'model/ForumModel.class.php';
