+++ lib/model/entity/agenda/AgendaItem.class.php
